Renamed evaluator model view to data_source_model_view
